Fix bug package web view
