+++ app/src/main/AndroidManifest.xml
@@ -20,8 +20,7 @@
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name" >
-        <activity
-            android:name=".CameraActivity"
+        <activity android:name=".CameraActivity"
             android:label="@string/app_name"
             android:screenOrientation="landscape" >
             <intent-filter>
@@ -30,8 +29,7 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity
-            android:name=".SettingsActivity"
+        <activity android:name=".SettingsActivity"
             android:label="@string/title_activity_settings"
             android:parentActivityName=".CameraActivity" >
             <meta-data
@@ -41,10 +39,6 @@
         <activity android:name=".listviewactivity.ListViewActivity" >
         </activity>
         <activity android:name=".statistics.StatisticsActivity" >
-        </activity>
-        <activity android:name=".statistics.webview.ChartDailyAverage" >
-        </activity>
-        <activity android:name=".statistics.webview.ChartNumberCrisisPerDay" >
         </activity>
         <activity android:name=".statistics.webview.ChartNumberCrisisPerMonth" >
         </activity>
