+++ app/src/main/java/com/hsbsoftwares/android/app/healthdiagnostic/statistics/webview/ChartNumberCrisisPerYear.java
@@ -47,7 +47,7 @@ public class ChartNumberCrisisPerYear extends Activity implements
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_cnumber_crisis_per_day);
+        setContentView(R.layout.activity_chart_number_crisis);
 
         //dateFormatter = new SimpleDateFormat("dd-MM-yyyy", Locale.getDefault());
         dateFormatter = new SimpleDateFormat("yyyy-MM-dd", Locale.getDefault());
@@ -60,9 +60,6 @@ public class ChartNumberCrisisPerYear extends Activity implements
         webView.getSettings().setJavaScriptEnabled(true);
         WebSettings webSettings = webView.getSettings();
         webSettings.setBuiltInZoomControls(true);
-
-
-
 
         setupActionBar();
     }
@@ -124,7 +121,7 @@ public class ChartNumberCrisisPerYear extends Activity implements
         html.append("google.setOnLoadCallback(drawChart);");
         html.append("function drawChart() {");
         html.append("var data = google.visualization.arrayToDataTable([");
-        html.append("['Year', 'Number Crisi'],");
+        html.append("['Year', 'Number Crisis'],");
         for (NumberCrisisPerYear ncpy  : numberCrisisPerYear){
             html.append("['");
             html.append(ncpy.getyear());
